% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/src_bigobject_sql.R
\name{src_bigobject}
\alias{src_bigobject}
\title{Connect to BigObject}
\usage{
src_bigobject(ip, port)
}
\arguments{
\item{ip}{string. The ip address of BigObject}

\item{port}{string. The port number of BigObject}
}
\value{
A BigObject src
}
\description{
Connect to BigObject
}

